[Ref] Refactor system initialization tasks into separate beans and use SystemInitializer to trigger them on startup.
